Gettin there
Included line of code that will detect gameover/gameclear so the program will stop. Bot has been improved to handle coming off of ladders specifically when the bot continues to move off the same way it came on the ladder. Bot has also indirectly been improved by slightly altering game_core.py code so that when an item is eaten, that cell is now considered empty so that the ai's v values will update based on items left on the map(uneaten). These improvements now allow the tanuki to complete the second task.