Getting Jumpy
Bot can now jump over needles and pitfalls that it deams safe to cross. Can nearly complete the third task. Has trouble getting stuck on two block platform with single fruit on the left and it also has trouble collecting the fruit to the right of the needle in the upper left corner